+++ client/web.js
@@ -1,0 +1,6 @@
+/**
+ * @param {import('vscode').ExtensionContext} ctx
+ */
+exports.activate = async function (ctx) {};
+
+exports.deactivate = function () {};