+++ webpack.config.js
@@ -1,0 +1,44 @@
+const {resolve} = require('path');
+const {ProvidePlugin} = require('webpack');
+
+/** @type {import('webpack').Configuration} */
+const webExtensionConfig = {
+	mode: 'none', // this leaves the source code as close as possible to the original (when packaging we set this to 'production')
+	target: 'webworker', // extensions run in a webworker context
+	entry: {
+		web: './client/web.js', // source of the web extension main file
+	},
+	output: {
+		filename: '[name].js',
+		path: resolve(__dirname, 'dist'),
+		libraryTarget: 'commonjs',
+	},
+	resolve: {
+		mainFields: ['browser', 'module', 'main'], // look for `browser` entry point in imported node modules
+		extensions: ['.js'], // support ts-files and js-files
+		alias: {
+			// provides alternate implementation for node module and source files
+		},
+		fallback: {
+			// Webpack 5 no longer polyfills Node.js core modules automatically.
+			// see https://webpack.js.org/configuration/resolve/#resolvefallback
+			// for the list of Node.js core module polyfills.
+			path: false,
+			fs: false,
+		},
+	},
+	plugins: [
+		new ProvidePlugin({
+			process: 'process/browser', // provide a shim for the global `process` variable
+		}),
+	],
+	externals: {
+		vscode: 'commonjs vscode', // ignored because it doesn't exist
+	},
+	performance: {
+		hints: false,
+	},
+	devtool: 'nosources-source-map', // create a source map that points to the original source file
+};
+
+module.exports = [webExtensionConfig];