+++ client/extension.js
@@ -1,6 +1,6 @@
-const {join} = require('path');
+// const {join} = require('path');
 const {initClient} = require('./lsp');
-const highlight = require('./highlight');
+// const highlight = require('./highlight');
 
 let client;
 
@@ -10,18 +10,18 @@ let client;
 exports.activate = async function (ctx) {
 	await Promise.all([
 		initClient(ctx).then(c => (client = c)),
-		highlight.init({
-			locateFamilyMarkupWasm(name) {
-				return ctx.asAbsolutePath(join('node_modules', 'highlight-familymarkup', name));
-			}
-		}),
+		// highlight.init({
+		// 	locateFamilyMarkupWasm(name) {
+		// 		return ctx.asAbsolutePath(join('node_modules', 'highlight-familymarkup', name));
+		// 	}
+		// }),
 	]);
 
-	return {
-		extendMarkdownIt(md) {
-			return md.use(highlight);
-		}
-	};
+	// return {
+	// 	extendMarkdownIt(md) {
+	// 		return md.use(highlight);
+	// 	}
+	// };
 };
 
 exports.deactivate = function () {